+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [unreleased]
+
+## [0.3.0]
 * Add `u.freeze` to freeze an object deeply. (https://github.com/substantial/updeep/issues/7)
 * Add `u.in` to update a single value in an object with a specified path. (https://github.com/substantial/updeep/issues/6)
 
@@ -33,7 +35,8 @@
 ## 0.1.0
 * Initial release
 
-[unreleased]: https://github.com/aaronjensen/updeep/compare/v0.2.3...HEAD
+[unreleased]: https://github.com/aaronjensen/updeep/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/aaronjensen/updeep/compare/v0.2.3...v0.3.0
 [0.2.3]: https://github.com/aaronjensen/updeep/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/aaronjensen/updeep/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/aaronjensen/updeep/compare/v0.2.0...v0.2.1